requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==45.0.5
Django==5.2.3
djangorestframework==3.16.0
ecdsa==0.19.1
jwt==1.4.0
Markdown==3.8.2
pyasn1==0.6.1
pycparser==2.22
python-decouple==3.8
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
